+++ app/src/main/java/com/trevjonez/polyadapter/sample/data/CategoryTitle.kt
@@ -1,5 +1,0 @@
-package com.trevjonez.polyadapter.sample.data
-
-data class CategoryTitle(
-    val text: CharSequence
-)