+++ app/src/main/java/com/trevjonez/polyadapter/sample/data/Movie.kt
@@ -1,7 +1,0 @@
-package com.trevjonez.polyadapter.sample.data
-
-data class Movie(
-    val title: CharSequence,
-    val url: String,
-    val imgUrl: String
-)