+++ app/src/main/java/polyadapter/sample/AppComponent.kt
@@ -1,0 +1,15 @@
+package polyadapter.sample
+
+import dagger.Component
+import dagger.android.AndroidInjectionModule
+import dagger.android.AndroidInjector
+import javax.inject.Singleton
+
+@Singleton
+@Component(
+    modules = [
+      SampleActivity.BindingModule::class,
+      AndroidInjectionModule::class
+    ]
+)
+interface AppComponent : AndroidInjector<App>