+++ app/src/main/java/polyadapter/sample/SampleGlideModule.kt
@@ -1,0 +1,7 @@
+package polyadapter.sample
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class SampleGlideModule : AppGlideModule()