+++ app/src/main/java/polyadapter/sample/data/CategoryTitle.kt
@@ -1,0 +1,5 @@
+package polyadapter.sample.data
+
+data class CategoryTitle(
+    val text: CharSequence
+)