+++ app/src/main/java/polyadapter/sample/viewholder/DataboundViewHolder.kt
@@ -1,0 +1,7 @@
+package polyadapter.sample.viewholder
+
+import androidx.databinding.ViewDataBinding
+import androidx.recyclerview.widget.RecyclerView
+
+class DataboundViewHolder<VDB : ViewDataBinding>(val viewBinding: VDB) :
+    RecyclerView.ViewHolder(viewBinding.root)