+++ app/src/main/java/polyadapter/sample/viewholder/DividerHolder.kt
@@ -1,0 +1,6 @@
+package polyadapter.sample.viewholder
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class DividerHolder(itemView: View) : RecyclerView.ViewHolder(itemView)